Move out GitLabCredentialMatcher to new package
