+++ src/main/java/io/jenkins/plugins/gitlabserverconfig/credentials/helpers/GitLabCredentialMatcher.java
@@ -1,8 +1,9 @@
-package io.jenkins.plugins.gitlabserverconfig.credentials;
+package io.jenkins.plugins.gitlabserverconfig.credentials.helpers;
 
 import com.cloudbees.plugins.credentials.Credentials;
 import com.cloudbees.plugins.credentials.CredentialsMatcher;
 import edu.umd.cs.findbugs.annotations.NonNull;
+import io.jenkins.plugins.gitlabserverconfig.credentials.PersonalAccessToken;
 import org.jenkinsci.plugins.plaincredentials.StringCredentials;
 
 public class GitLabCredentialMatcher implements CredentialsMatcher {
